+++ tests/test_open_gzipped.py
@@ -4,14 +4,21 @@
 """Tests for `seamm_util` package."""
 
 import seamm_util
-import os
-import os.path
+from pathlib import Path
+
+datapath = Path(__file__).parent / "data"
+
+
+def uri_handler(path):
+    if str(path).startswith("data:"):
+        return (datapath / str(path)[5:]).expanduser().resolve()
+    else:
+        return Path(path)
 
 
 def test_open_1file():
     """Testing that we can read a simple file"""
-    datapath = os.path.join(os.path.dirname(__file__), "data")
-    filepath = os.path.join(datapath, "file1.txt.gz")
+    filepath = datapath / "file1.txt.gz"
     data = ["file1 line 1", "file1 line 2", "file1 line 3", "file1 line 4"]
 
     i = 0
@@ -23,8 +30,7 @@ def test_open_1file():
 
 def test_open_2files():
     """Testing that the default 'include' works"""
-    datapath = os.path.join(os.path.dirname(__file__), "data")
-    filepath = os.path.join(datapath, "file_middle.txt.gz")
+    filepath = datapath / "file_middle.txt.gz"
     data = [
         "file_middle line 1",
         "file_middle line 2",
@@ -42,10 +48,31 @@ def test_open_2files():
             i += 1
 
 
+def test_open_uri():
+    """Testing that the default 'include' works"""
+    filepath = "data:file_uri.txt.gz"
+    data = [
+        "file_uri line 1",
+        "file_uri line 2",
+        "file1 line 1",
+        "file1 line 2",
+        "file1 line 3",
+        "file1 line 4",
+        "file_uri line 3",
+    ]
+
+    i = 0
+    with seamm_util.Open(
+        filepath, "r", include="include", uri_handler=uri_handler
+    ) as fd:
+        for line in fd:
+            assert line.strip() == data[i]
+            i += 1
+
+
 def test_open_2files_at_end():
     """Testing that the default 'include' works at the end of a file"""
-    datapath = os.path.join(os.path.dirname(__file__), "data")
-    filepath = os.path.join(datapath, "file_end.txt.gz")
+    filepath = datapath / "file_end.txt.gz"
     data = [
         "file_end line 1",
         "file_end line 2",
@@ -68,8 +95,7 @@ def test_lines():
 
     Note that the line count includes the line with 'include'
     """
-    datapath = os.path.join(os.path.dirname(__file__), "data")
-    filepath = os.path.join(datapath, "file_middle.txt.gz")
+    filepath = datapath / "file_middle.txt.gz"
 
     num = [1, 2, 1, 2, 3, 4, 4]
     i = 0
@@ -84,8 +110,7 @@ def test_total_lines():
 
     Note that the line count includes the line with 'include'
     """
-    datapath = os.path.join(os.path.dirname(__file__), "data")
-    filepath = os.path.join(datapath, "file_middle.txt.gz")
+    filepath = datapath / "file_middle.txt.gz"
 
     with seamm_util.Open(filepath, "r", include="include") as fd:
         for line in fd:
@@ -95,8 +120,7 @@ def test_total_lines():
 
 def test_nested_includes():
     """Testing an include within an include"""
-    datapath = os.path.join(os.path.dirname(__file__), "data")
-    filepath = os.path.join(datapath, "file_include1.txt.gz")
+    filepath = datapath / "file_include1.txt.gz"
     data = [
         "file_include1 line1",
         "file_end line 1",
@@ -117,8 +141,7 @@ def test_nested_includes():
 
 def test_include_empty_file():
     """Testing that including an empty file works"""
-    datapath = os.path.join(os.path.dirname(__file__), "data")
-    filepath = os.path.join(datapath, "include_empty_file.txt.gz")
+    filepath = datapath / "include_empty_file.txt.gz"
     data = ["file_end line 1", "file_end line 2", "file_end line 3"]
 
     i = 0
@@ -132,8 +155,7 @@ def test_blank_lines():
     """Testing that empty lines and ones with just blanks don't
     cause any problems
     """
-    datapath = os.path.join(os.path.dirname(__file__), "data")
-    filepath = os.path.join(datapath, "include_blank_lines.txt.gz")
+    filepath = datapath / "include_blank_lines.txt.gz"
     data = [
         "file_end line 1",
         "",
@@ -155,12 +177,12 @@ def test_blank_lines():
 
 def test_stack():
     """Testing the stack of files part way through reading includes"""
-    cwd = os.path.dirname(__file__)
-    datapath = os.path.join(cwd, "data")
-    filepath = os.path.join(datapath, "file_include1.txt.gz")
+    cwd = Path(__file__).parent
+    filepath = datapath / "file_include1.txt.gz"
     data = [
-        os.path.normcase("data/file1.txt.gz:2"),
-        os.path.normcase("data/file_end.txt.gz:4"),
+        "data/file1.txt.gz:2",
+        "data/file_end.txt.gz:4",
+        "data/file_include1.txt.gz:2",
     ]
 
     i = 0
@@ -170,15 +192,18 @@ def test_stack():
             if i == 6:
                 stack = []
                 for line in fd.stack():
-                    stack.append(os.path.relpath(line, cwd))
+                    stack.append(str(Path(line).relative_to(cwd)).lower())
+                print("Stack")
+                print("\t" + "\n\t".join(stack))
+                print("should be")
+                print("\t" + "\n\t".join(data))
                 assert stack == data
                 break
 
 
 def test_different_include():
     """Testing that '#include' works"""
-    datapath = os.path.join(os.path.dirname(__file__), "data")
-    filepath = os.path.join(datapath, "file_#include.txt.gz")
+    filepath = datapath / "file_#include.txt.gz"
     data = [
         "file_middle line 1",
         "file_middle line 2",
@@ -198,8 +223,7 @@ def test_different_include():
 
 def test_push():
     """Testing that the default 'include' works"""
-    datapath = os.path.join(os.path.dirname(__file__), "data")
-    filepath = os.path.join(datapath, "file_middle.txt.gz")
+    filepath = datapath / "file_middle.txt.gz"
     data = [
         "file_middle line 1",
         "file_middle line 2",